fixed problems caused by importing new code

+++ shcp-developer/src/main/java/com/shcp/developer/service/impl/ExamineServiceImpl.java
@@ -1,9 +1,12 @@
 package com.shcp.developer.service.impl;
 
 import com.shcp.common.pojo.ShcpResult;
+import com.shcp.dao.mapper.DeveloperinfoMapper;
 import com.shcp.developer.pojo.ExamineStatus;
 import com.shcp.developer.service.ExamineService;
 import com.shcp.developer.utils.IdGenerator;
+import com.shcp.pojo.Developerinfo;
+import com.shcp.pojo.TbDeveloper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -29,7 +32,7 @@ public class ExamineServiceImpl implements ExamineService{
         developerinfo.setDcstatus(ExamineStatus.AUDITING);
         developerinfo.setDcnumber(dcNumber);
         developerinfo.setDcid(IdGenerator.generateExamineId());
-        developerinfo.setDcaid(0);
+        developerinfo.setDcaid((long) 0);
         developerinfoMapper.insertSelective(developerinfo);
         log.info("developerinfo:{} insert success", developerinfo);
         return ShcpResult.ok();
